change roots, admin controllers patches and namespace

+++ routes/web.php
@@ -46,32 +46,32 @@ Route::middleware('auth')->group(function(){
 // Admin Panel
 
 Route::prefix('admin')->middleware('role:admin')->group(function(){
-    Route::get('/panel','AdminController@index')->name('admin-panel');
+    Route::get('/panel','Admin\AdminController@index')->name('admin-panel');
 
-    Route::get('/programs','AdminProgramController@index')->name('admin-programs-index');
-    Route::get('/programs/create','AdminProgramController@create')->name('admin-program-create');
-    Route::post('/programs/create','AdminProgramController@store')->name('admin-program-store');
-    Route::get('/programs/{program_id}','AdminProgramController@detail')->name('admin-program-detail');
-    Route::get('/programs/{program_id}/edit','AdminProgramController@edit')->name('admin-program-edit');
-    Route::post('/programs/{program_id}/edit','AdminProgramController@save')->name('admin-program-save');
-    Route::delete('/programs/{program_id}','AdminProgramController@delete')->name('admin-program-delete');
+    Route::get('/programs','Admin\AdminProgramController@index')->name('admin-programs-index');
+    Route::get('/programs/create','Admin\AdminProgramController@create')->name('admin-program-create');
+    Route::post('/programs/create','Admin\AdminProgramController@store')->name('admin-program-store');
+    Route::get('/programs/{program_id}','Admin\AdminProgramController@detail')->name('admin-program-detail');
+    Route::get('/programs/{program_id}/edit','Admin\AdminProgramController@edit')->name('admin-program-edit');
+    Route::post('/programs/{program_id}/edit','Admin\AdminProgramController@save')->name('admin-program-save');
+    Route::delete('/programs/{program_id}','Admin\AdminProgramController@delete')->name('admin-program-delete');
 
-    Route::get('/users','AdminUserController@index')->name('admin-users-index');
-    Route::get('/users/{user_id}','AdminUserController@detail')->name('admin-user-detail');
-    Route::delete('/users/{user_id}','AdminUserController@delete')->name('admin-user-delete');
+    Route::get('/users','Admin\AdminUserController@index')->name('admin-users-index');
+    Route::get('/users/{user_id}','Admin\AdminUserController@detail')->name('admin-user-detail');
+    Route::delete('/users/{user_id}','Admin\AdminUserController@delete')->name('admin-user-delete');
 
-    Route::get('/documents','AdminDocumentController@index')->name('admin-documents-index');
-    Route::get('/documents/upload','AdminDocumentController@uploadForm')->name('admin-document-upload');
-    Route::post('/documents/upload','AdminDocumentController@store')->name('admin-document-store');
-    Route::delete('/documents/{document_id}','AdminDocumentController@delete')->name('admin-document-delete');
+    Route::get('/documents','Admin\AdminDocumentController@index')->name('admin-documents-index');
+    Route::get('/documents/upload','Admin\AdminDocumentController@uploadForm')->name('admin-document-upload');
+    Route::post('/documents/upload','Admin\AdminDocumentController@store')->name('admin-document-store');
+    Route::delete('/documents/{document_id}','Admin\AdminDocumentController@delete')->name('admin-document-delete');
 
-    Route::get('/reviews','AdminReviewController@index')->name('admin-requests-index');
-    Route::get('/reviews/{review_id}','AdminReviewController@edit')->name('admin-request-detail');
-    Route::get('/reviews/create','AdminReviewController@create')->name('admin-request-create');
-    Route::post('/reviews/create','AdminReviewController@store')->name('admin-request-store');
-    Route::get('/reviews/{review_id}/edit','AdminReviewController@edit')->name('admin-request-edit');
-    Route::post('/reviews/{review_id}/edit','AdminReviewController@save')->name('admin-request-save');
-    Route::delete('/reviews/{review_id}','AdminReviewController@delete')->name('admin-request-delete');
+    Route::get('/reviews','Admin\AdminReviewController@index')->name('admin-requests-index');
+    Route::get('/reviews/{review_id}','Admin\AdminReviewController@edit')->name('admin-request-detail');
+    Route::get('/reviews/create','Admin\AdminReviewController@create')->name('admin-request-create');
+    Route::post('/reviews/create','Admin\AdminReviewController@store')->name('admin-request-store');
+    Route::get('/reviews/{review_id}/edit','Admin\AdminReviewController@edit')->name('admin-request-edit');
+    Route::post('/reviews/{review_id}/edit','Admin\AdminReviewController@save')->name('admin-request-save');
+    Route::delete('/reviews/{review_id}','Admin\AdminReviewController@delete')->name('admin-request-delete');
 });
 
 require __DIR__.'/auth.php';
